+++ CHANGES.rst
@@ -18,6 +18,7 @@ assign_wcs
 
 associations
 ------------
+- Exclude mir_lrs-slitless calibration data from level 3 processing [#3990]
 
 - Fix in load_as_asn for UTF-8 errors [#3942]